% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ukaq_daqi.R
\name{import_ukaq_daqi}
\alias{import_ukaq_daqi}
\title{Import UK Daily Air Quality Index Data}
\usage{
import_ukaq_daqi(
  year,
  code = NULL,
  source = "ukaq",
  pollutant = c("no2", "o3", "pm10", "pm2.5", "so2"),
  append_metadata = FALSE,
  metadata_columns = c("site_type", "latitude", "longitude"),
  pivot = "long",
  ...,
  .return = NULL
)
}
\arguments{
\item{year}{\emph{A year, or range of years, from which to import data.}

\strong{required}

\code{year} expects a vector of integer values. It can be a single year
(\code{2020}), a selection of specific years (c(\verb{2020, 2021, 2022})), or a range
of years (\code{2020:2025}).}

\item{code}{\emph{Specific site codes to import.}

\emph{default:} \code{NA}

When importing DAQI data, the default behaviour is for all sites in a given
network to be imported. Specifying \code{code} will filter the result for
specific sites based on their site codes, available through
\code{\link[=import_ukaq_meta]{import_ukaq_meta()}}.}

\item{source}{\emph{One or more UK Monitoring networks from which to import data.}

\emph{default:} \code{"ukaq"}

The default, \code{"ukaq"}, will import data from the AURN, the four 'devolved'
networks, and locally managed English monitoring networks. Any combination
of \code{"aurn"}, \code{"aqe"}, \code{"saqn"}, \code{"waqn"}, and \code{"niaqn"} will
\emph{only} import data from those specific monitoring networks. Note that a
mismatch between \code{code} and \code{source} may result in no data being imported.}

\item{pollutant}{\emph{One or more DAQI pollutants for which to import data.}

\emph{default:} no2, o3, pm10, pm2.5, so2

By default, all 5 DAQI pollutants are
imported by \code{\link[=import_ukaq_daqi]{import_ukaq_daqi()}}. \code{pollutant} allows any specific
combination of the DAQI pollutants to be returned instead.}

\item{append_metadata}{\emph{Append site metadata to the dataframe?}

\emph{default:} \code{FALSE}

When \code{TRUE}, the resulting \code{data.frame} will have site metadata appended.
The specific columns are selected using \code{metadata_columns}.}

\item{metadata_columns}{\emph{Specific metadata columns to append to the
data.frame.}

\emph{default:} \code{c("site_type", "latitude", "longitude")}

When \code{append_metadata = TRUE}, the columns selected here will be appended
to the data. Columns names should match those in \code{\link[=import_ukaq_meta]{import_ukaq_meta()}},
excluding pollutant-specific columns (e.g., ratification date).}

\item{pivot}{\emph{Should the dataframe be 'wide' or 'long'?}

\emph{default:} \code{"long"}

There are two main ways to store air quality data; "long" (with 'pollutant'
and 'value' columns) or "wide" (with each pollutant value being stored in
its own column). \code{pivot} allows users to define which format they would
prefer their data in.}

\item{...}{\emph{Not used.}}

\item{.return}{\emph{Signifier for the dataframe class.}

\emph{default:} \code{NULL}

\link[=ukaq-package]{ukaq} functions, by default, will return \code{tbl_df}s if the
\link[tibble:tibble-package]{tibble} package is installed, but will otherwise
return \code{data.frame}s. \code{.return} can override this behaviour, and takes
either \code{"tbl"} or \code{"df"}, which sets the return class to be \code{tbl_df} or
\code{data.frame} respectively.}
}
\value{
a \code{data.frame}
}
\description{
Imports \href{https://uk-air.defra.gov.uk/air-pollution/daqi}{UK Daily Air Quality Index (DAQI)} data for selected
pollutants, years, and sites. Returned information includes the relevant
daily statistic (e.g., daily average for PM10), the index (1:10) and the band
(Low, Moderate, High, or Very High).
}
\references{
With thanks to Trevor Davies and \href{https://www.ricardo.com/en}{Ricardo Plc} for preparing and hosting the data
}
\seealso{
\url{https://uk-air.defra.gov.uk/air-pollution/daqi} for more
information about the DAQI.
}
\author{
Jack Davison, David Carslaw
}
