% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ukaq_meta.R
\name{import_ukaq_meta}
\alias{import_ukaq_meta}
\alias{import_ukaq_pollutants}
\title{Import UK Air Quality Metadata}
\usage{
import_ukaq_meta(
  source = "ukaq",
  year = NA,
  by_pollutant = FALSE,
  ...,
  .return = NULL
)

import_ukaq_pollutants(source = "ukaq", ..., .return = NULL)
}
\arguments{
\item{source}{\emph{One or more UK Monitoring networks from which to import
metadata}

\emph{default:} \code{"ukaq"}

The default, \code{"ukaq"}, will import metadata from the AURN, the four
'devolved' networks, and locally managed English monitoring networks. Any
combination of \code{"aurn"}, \code{"aqe"}, \code{"saqn"}, \code{"waqn"}, \code{"niaqn"} and ,
\code{"lmam"} will \emph{only} import data from those specific monitoring networks.}

\item{year}{\emph{A year, or range of years, with which to filter data.}

\emph{default:} \code{NA}

By default all monitoring stations are returned regardless of their
open/closed status. \code{year} allows users to filter sites open in a specific
year, or over a range of years. For example, \code{year = c(2010, 2020)} (or
\code{year = 2010:2020}) will return sites that were open between 2010 and 2020.}

\item{by_pollutant}{\emph{Return site-pollutant combinations?}

\emph{default:} \code{FALSE}

If \code{TRUE}, the rows of the resulting \code{data.frame} will each represent a
monitoring site-pollutant pair, with the start, end, and ratification date
reflecting each specific pollutant. When \code{FALSE}, the default, each row is
a unique monitoring site, and the start and end dates represent the opening
and closing dates of the site as a whole.}

\item{...}{\emph{Not used.}}

\item{.return}{\emph{Signifier for the dataframe class.}

\emph{default:} \code{NULL}

\link[=ukaq-package]{ukaq} functions, by default, will return \code{tbl_df}s if the
\link[tibble:tibble-package]{tibble} package is installed, but will otherwise
return \code{data.frame}s. \code{.return} can override this behaviour, and takes
either \code{"tbl"} or \code{"df"}, which sets the return class to be \code{tbl_df} or
\code{data.frame} respectively.}
}
\value{
a \code{data.frame}
}
\description{
\code{\link[=import_ukaq_meta]{import_ukaq_meta()}} imports metadata for UK air quality monitoring sites. By
default this function returns one row per monitoring site, including the site
opening and closing dates and other useful meta-data. The \code{by_pollutant}
option will return one row per monitoring site \emph{and pollutant}; see
\code{\link[=import_ukaq_pollutants]{import_ukaq_pollutants()}} for definitions. The \code{"site"} column can be used
in other functions to import data.
}
\references{
With thanks to Trevor Davies and \href{https://www.ricardo.com/en}{Ricardo Plc} for preparing and hosting the data
}
\author{
Jack Davison, David Carslaw
}
