% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{palette_daqi}
\alias{palette_daqi}
\alias{palette_gaf}
\title{Get palettes of colours related to UK Air Quality legislation}
\usage{
palette_daqi(option = c("index", "bands"), named = FALSE)

palette_gaf(
  option = c("categorical", "duo", "focus", "sequential"),
  named = FALSE
)
}
\arguments{
\item{option}{\emph{The palette of interest.}

One of the following palette options:
\itemize{
\item \strong{daqi:} \code{"index"} (the default) or \code{"band"}
\item \strong{gaf:} \code{"categorical"} (the default), \code{"duo"}, \code{"focus"}, or \code{"sequential"}
}}

\item{named}{\emph{Return a named vector?}

Defaults to \code{FALSE}, which will return an unnamed vector. Named vectors may
be of interest if you need to index specific colours rather than use the
whole palette. The specific names vary with the specific function and
\code{option}.}
}
\description{
These functions are convenient ways to obtain vectors of colours for charts
and tables. \link{palette_daqi} has options \code{"index"} or \code{"bands"} and returns the
colours used for the DAQI on \href{https://uk-air.defra.gov.uk/air-pollution/daqi}{UK AIR}. \link{palette_gaf} has
options \code{"categorical"}, \code{"duo"}, \code{"focus"}, and \code{"sequential"} and returns
the suggested colour palettes by the \href{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}{UK Government Analysis Function}.
}
\seealso{
\url{https://uk-air.defra.gov.uk/air-pollution/daqi}

\url{https://analysisfunction.civilservice.gov.uk/policy-store/data-visualisation-colours-in-charts/}
}
