% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukaq-package.R
\docType{package}
\name{ukaq-package}
\alias{ukaq}
\alias{ukaq-package}
\title{ukaq: Access UK Air Quality Monitoring Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Importer for a myriad of United Kingdom air quality data, including monitoring station metadata, hourly measured concentrations, and pre-calculated statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openair-project/ukaq}
  \item \url{https://openair-project.github.io/ukaq/}
  \item Report bugs at \url{https://github.com/openair-project/ukaq/issues}
}

}
\author{
\strong{Maintainer}: Jack Davison \email{jack.davison@ricardo.com} (\href{https://orcid.org/0000-0003-2653-6615}{ORCID})

}
\keyword{internal}
