% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ukaq_daqi.R
\name{import_ukaq_daqi}
\alias{import_ukaq_daqi}
\title{Import UK Daily Air Quality Index Data}
\usage{
import_ukaq_daqi(
  code = NULL,
  year,
  source = "aurn",
  pollutant = NULL,
  append_metadata = FALSE,
  metadata_columns = c("site_type", "latitude", "longitude"),
  pivot = "long",
  progress = NA,
  ...,
  .class = NULL
)
}
\arguments{
\item{code}{\emph{Specific site codes to import.}

\emph{default:} \code{NULL}

\code{code} expects a vector of character values. It defines the specific sites
to import will filter the result for specific sites based on their site
codes, available through \code{\link[=import_ukaq_meta]{import_ukaq_meta()}}. Note that a mismatch between
\code{code} and \code{source} may result in no data being imported. If \code{NULL}, the
default, all available data for the \code{year} and \code{source} will be returned.}

\item{year}{\emph{A year, or range of years, from which to import data.}

\strong{required}

\code{year} expects a vector of integer values. It can be a single year
(\code{2020}), a selection of specific years (c(\verb{2020, 2021, 2022})), or a range
of years (\code{2020:2025}).}

\item{source}{\emph{One or more UK Monitoring networks from which to import data.}

\emph{default:} \code{"ukaq"}

The default, \code{"ukaq"}, will look for sites from any of the available
networks. Any combination of \code{"aurn"}, \code{"aqe"}, \code{"saqn"}, \code{"waqn"}, or
\code{"niaqn"} will \emph{only} import data from those specific monitoring networks.
Note that a mismatch between \code{code} and \code{source} may result in no data
being imported. The DAQI is not pre-calculated for \code{"lmam"} sites.}

\item{pollutant}{\emph{One or more DAQI pollutants for which to import data.}

\emph{default:} \code{NULL}

By default, all available pollutants are imported. \code{pollutant} allows any
specific combination of the pollutants to be returned instead. Note that
this should be a DAQI pollutant - i.e., one or more of \code{"no2"}, \code{"pm10"},
\code{"pm2.5"}, \code{"o3"} or \code{"so2"}.}

\item{append_metadata}{\emph{Append site metadata to the dataframe?}

\emph{default:} \code{FALSE}

When \code{TRUE}, the resulting \code{data.frame} will have site metadata appended.
The specific columns are selected using \code{metadata_columns}.}

\item{metadata_columns}{\emph{Specific metadata columns to append to the
data.frame.}

\emph{default:} \code{c("site_type", "latitude", "longitude")}

When \code{append_metadata = TRUE}, the columns selected here will be appended
to the data. Columns names should match those in \code{\link[=import_ukaq_meta]{import_ukaq_meta()}},
excluding pollutant-specific columns (e.g., ratification date).}

\item{pivot}{\emph{Should the dataframe be 'wide' or 'long'?}

\emph{default:} \code{"wide"}

There are two main ways to store air quality data; "long" (with 'pollutant'
and 'value' columns) or "wide" (with each pollutant value being stored in
its own column). \code{pivot} allows users to define which format they would
prefer their data in.}

\item{progress}{\emph{Show a progress bar?}

\emph{default:} \code{NA}

When \code{TRUE}, this function will print a progress bar to track individual
files being imported. If \code{FALSE} this is supressed. If \code{NA}, the default,
the function will work out if a progress bar would be useful (i.e., if the
function is being run in an interactive session and if more than one remote
file is being accessed).}

\item{...}{\emph{Not used.}}

\item{.class}{\emph{Signifier for the dataframe class.}

\emph{default:} \code{NULL}

\link[=ukaq-package]{ukaq} functions, by default, will return \code{tbl_df}s if the
\link[tibble:tibble-package]{tibble} package is installed, but will otherwise
return \code{data.frame}s. \code{.class} can override this behaviour, and takes
either \code{"tbl"} or \code{"df"}, which sets the return class to be \code{tbl_df} or
\code{data.frame} respectively.}
}
\value{
a \code{data.frame}
}
\description{
Imports \href{https://uk-air.defra.gov.uk/air-pollution/daqi}{UK Daily Air Quality Index (DAQI)} data for selected
pollutants, years, and sites. Returned information includes the relevant
daily statistic (e.g., daily average for PM10), the index (1:10) and the band
(Low, Moderate, High, or Very High).
}
\references{
With thanks to Trevor Davies and \href{https://www.ricardo.com/en}{Ricardo Plc} for preparing and hosting the data
}
\seealso{
\url{https://uk-air.defra.gov.uk/air-pollution/daqi} for more
information about the DAQI.
}
\author{
Jack Davison, David Carslaw
}
