% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukaq-package.R
\docType{package}
\name{ukaq-package}
\alias{ukaq}
\alias{ukaq-package}
\title{ukaq: Access UK Air Quality Monitoring Data}
\description{
Importer for a myriad of United Kingdom air quality data, including monitoring station metadata, hourly measured concentrations, and pre-calculated statistics.
}
\author{
\strong{Maintainer}: Jack Davison \email{jack.davison@ricardo.com}

}
\keyword{internal}
